change 'dashboard' to 'userProfile'
