+++ app/app.js
@@ -10,7 +10,8 @@ var loginRouter = require('./routes/login');
 var signupRouter = require('./routes/signup');
 var signupSubmitRouter = require('./routes/signupSubmit');
 var loginSubmitRouter = require('./routes/loginSubmit');
-var dashboardRouter = require('./routes/dashboard');
+var dashboardRouter = require('./routes/dashboards');
+var userProfileRouter = require('./routes/userProfile');
 var errorLoginRouter = require('./routes/error_login');
 var deleteSubmitRouter = require('./routes/deleteSubmit');
 var editSubmitRouter = require('./routes/editSubmit');
@@ -47,6 +48,7 @@ app.use('/signup', signupRouter);
 app.use('/signupSubmit', signupSubmitRouter.router);
 app.use('/loginSubmit', loginSubmitRouter);
 app.use('/dashboard', dashboardRouter);
+app.use('/userProfile', userProfileRouter);
 app.use('/error_login', errorLoginRouter);
 app.use('/deleteSubmit', deleteSubmitRouter);
 app.use('/editSubmit', editSubmitRouter);
