--- app/views/dashboard.jade
@@ -1,75 +1,0 @@
-doctype html 
-html 
-    head 
-        meta(charset="utf-8")
-        meta(name="viewport", content="width=device-width, initial-scale=1, shrink-to-fit=no")
-        link(rel="stylesheet", href="https://fonts.googleapis.com/css?family=Roboto:400,700")
-        title Dashboard 
-        link(rel="stylesheet", href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css")
-        script(src="https://code.jquery.com/jquery-3.5.1.min.js")
-        script(src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js")
-        script(src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js")
-        link(rel="stylesheet", href="stylesheets/dashboard.css")
-
-body 
-    div(class="container")
-        div(class="row gutters")
-
-            div(class="col-xl-3 col-lg-3 col-md-12 col-sm-12 col-12")
-                div(class="card h-100")
-                    div(class="card-body")
-                        div(class="account-settings")
-                            div(class="user-profile")
-                                div(class="user-avatar")
-                                    img(src="http://www.newdesignfile.com/postpic/2014/02/windows-8-user-account-icon_325746.png", id="icon", alt="User Icon")
-                                h5(class="user-name") #{userInfo.first_name} #{userInfo.last_name}
-                                h6(class="user-email") #{userInfo.email}
-                    div(class="card-footer")
-                        div(class="text-center")
-                            form(method="get" action="/")
-                                button(type="submit" id="log-out-submit" name="log-out-submit" class="btn btn-secondary") Log Out
-
-            div(class="col-xl-9 col-lg-9 col-md-12 col-sm-12 col-12")
-                div(class="card h-100")
-                    div(class="card-body")
-                        form(method="post" action="/editSubmit")
-                            div(class="row gutters")
-                                div(class="col-xl-12 col-lg-12 col-md-12 col-sm-12 col-12")
-                                    h6(class="mb-2 text-primary") Edit Profile
-                                div(class="col-xl-6 col-lg-6 col-md-6 col-sm-6 col-12")
-                                    div(class="form-group")
-                                        label(for="firstName") First Name
-                                        input(type="text" class="form-control" name="firstName" placeholder="Enter first name")
-                                div(class="col-xl-6 col-lg-6 col-md-6 col-sm-6 col-12")
-                                    div(class="form-group")
-                                        label(for="newEmail") Email
-                                        input(type="email" class="form-control" name="newEmail" placeholder="Enter email")
-                                div(class="col-xl-6 col-lg-6 col-md-6 col-sm-6 col-12")
-                                    div(class="form-group")
-                                        label(for="lastName") Last Name
-                                        input(type="text" class="form-control" name="lastName" placeholder="Enter last name")
-                            div(class="row gutters")
-                                div(class="col-xl-6 col-lg-6 col-md-6 col-sm-6 col-12")
-                                    div(class="form-group")
-                                        label(for="oldPassword") Old Password
-                                        input(type="password" class="form-control" name="oldPassword" placeholder="Enter old password")
-                                div(class="col-xl-6 col-lg-6 col-md-6 col-sm-6 col-12")
-                                    div(class="form-group")
-                                        label(for="newPassword") New Password
-                                        input(type="password" class="form-control" name="newPassword" placeholder="Enter new password")
-                                div(class="col-xl-6 col-lg-6 col-md-6 col-sm-6 col-12")
-                                    div(class="form-group")
-                                        label(for="confirmNewPassword") Confirm New Password
-                                        input(type="password" class="form-control" name="confirmNewPassword" placeholder="Re-enter new password")
-                            div(class="row gutters")
-                                div(class="col-xl-12 col-lg-12 col-md-12 col-sm-12 col-12")
-                                    div(class="text-right")
-                                        input(type="hidden" name="oldEmail" value="#{userInfo.email}")
-                                        input(type="hidden" name="validOldPassword" value="#{userInfo.password}")
-                                        button(type="submit" id="submit" name="submit" class="btn btn-primary") Update
-                                        
-                    div(class="card-footer")
-                        form(method="post" action="/deleteSubmit")
-                            input(type="hidden" id="email" name="email" value="#{userInfo.email}")
-                            button(type="submit" id="delete-submit" name="delete-submit" class="btn btn-secondary") Delete Profile
-
